+++ test2.0/e2e/networks/ttn.js
@@ -1,9 +1,10 @@
 const Client = require('../../../rest/networkProtocols/handlers/TheThingsNetwork/TtnRestClient')
 const client = new Client()
 const { key } = require('ttn')
+const crypto = require('crypto')
 
-const timestamp = Date.now()
-const APP_ID = `lpwansvr-e2e-app-${timestamp}`
+const randomString = crypto.randomBytes(6).toString('base64').toLowerCase()
+const APP_ID = `lpwansvr-e2e-app-hmoxiloq`
 
 module.exports = {
   client,
@@ -43,12 +44,12 @@ module.exports = {
     altitude: 0,
     app_id: APP_ID,
     description: 'CableLabs TTN Device ABP',
-    dev_id: `lpwansvr-e2e-abp-${timestamp}`,
+    dev_id: `lpwansvr-e2e-abp-hmoxiloq`,
     latitude: 52.375,
     longitude: 4.887,
     lorawan_device: {
-      dev_eui: '00A0A112727496D3',
-      dev_id: `lpwansvr-e2e-abp-${timestamp}`,
+      dev_eui: '006476D29AA36C97',
+      dev_id: `lpwansvr-e2e-abp-hmoxiloq`,
       app_id: APP_ID,
       activation_constraints: 'abp',
       app_s_key: key(16),
@@ -63,12 +64,12 @@ module.exports = {
     altitude: 0,
     appID: APP_ID,
     description: 'TTN Device Using OTAA',
-    dev_id: `lpwansvr-e2e-otaa-${timestamp}`,
+    dev_id: `lpwansvr-e2e-otaa-hmoxiloq`,
     latitude: 52.375,
     longitude: 4.887,
     lorawan_device: {
-      dev_eui: '0099F6D395BD932A',
-      dev_id: `lpwansvr-e2e-otaa-${timestamp}`,
+      dev_eui: '003C48905977AD1C',
+      dev_id: `lpwansvr-e2e-otaa-hmoxiloq`,
       app_id: APP_ID,
       activation_constraints: 'otaa',
       app_key: key(16)
